modularize contact and utils
