+++ backend/utils.js
@@ -1,0 +1,7 @@
+const re = /[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?/
+
+const validEmail = (email) => {
+  return re.test(email);
+}
+
+module.exports = validEmail;
